requirements.txt
# Core Framework
streamlit>=1.28.0

# PDF Processing
PyPDF2>=3.0.0
pypdf>=3.0.0
reportlab>=4.0.0
pikepdf>=8.0.0

# Image Processing
Pillow>=10.0.0
pdf2image>=1.16.0

# Document Conversion
python-docx>=1.0.0
openpyxl>=3.0.0
python-pptx>=0.6.0

# OCR
pytesseract>=0.3.0
pdf2image>=1.16.0

# PDF Security
cryptography>=41.0.0

# HTML to PDF
pdfkit>=1.0.0
weasyprint>=60.0

# Compression
img2pdf>=0.4.0

# Utilities
python-dotenv>=1.0.0
